+++ src/main/views/macro/form.njk
@@ -462,7 +462,7 @@
 {% endmacro %}
 
 {% macro apiErrorPostCodeField(error) %}
-  {% if error == 'Enter a valid postcode. Try again using another postcode.' %}
+  {% if error == 'Enter a postcode for England or Wales' %}
     <div class="form-group form-group-error">
       <label for="formfield" id="formfield[label]" class="form-label">{{ t(label) }}
       <span class="error-message">{{ t(error) }}</span>
@@ -471,7 +471,7 @@
 {% endmacro %}
 
 {% macro apiErrorCourtNameField(error) %}
-  {% if error == 'Could not find a court name to match. Try searching by postcode instead.' %}
+  {% if error == 'Enter a court or place name in England or Wales' %}
     <div class="form-group form-group-error">
       <label for="formfield" id="formfield[label]" class="form-label">{{ t(label) }}
       <span class="error-message">{{ t(error) }}</span>
